// src/app/shared/models/product.ts
export class Product {
  $key: string;
  productId: number;
  productName: string;
  productCategory: string;
  productPrice: number;
  productDescription: string;
  productAdded: number;
  productQuatity: number;
  cartQuantity: number;
  productQuantity: number;
  ratings: number;
  favourite: boolean;
  productSeller: string;
  productSKU: string;
  enable: boolean;
  images: [];
  bestProducts: boolean;
}